@@ -20,15 +20,15 @@ repositories {
 
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "com.chaemil.hgms"
         minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 2018041401
-        versionName "2.8.4"
+        targetSdkVersion 28
+        versionCode 2019010401
+        versionName "2.8.5"
 
         multiDexEnabled true
         renderscriptTargetApi 19
