@@ -1,41 +1,57 @@
 package com.chaemil.hgms.model;
 
-import com.novoda.downloadmanager.lib.DownloadManager;
+import android.app.DownloadManager;
+
+import com.orm.SugarRecord;
 
 /**
  * Created by chaemil on 30.6.16.
  */
-public class Download {
-    private final int downloadStatus;
-    private final long batchId;
-    private final long videoId;
-    private final String filename;
+public class Download extends SugarRecord {
+    private int downloadStatus;
+    private String batchId;
+    private long videoId;
+    private String filename;
 
-    public Download(long videoId, String filename, int downloadStatus, long batchId) {
+    public Download() {
+    }
+
+    public Download(long videoId, String filename, int downloadStatus, String batchId) {
         this.videoId = videoId;
         this.filename = filename;
         this.downloadStatus = downloadStatus;
         this.batchId = batchId;
     }
 
+    public int getDownloadStatus() {
+        return downloadStatus;
+    }
+
+    public void setDownloadStatus(int downloadStatus) {
+        this.downloadStatus = downloadStatus;
+    }
+
+    public String getBatchId() {
+        return batchId;
+    }
+
+    public void setBatchId(String batchId) {
+        this.batchId = batchId;
+    }
+
+    public long getVideoId() {
+        return videoId;
+    }
+
+    public void setVideoId(long videoId) {
+        this.videoId = videoId;
+    }
+
     public String getFilename() {
         return filename;
     }
 
-    public long getVideoServerId() {
-        return videoId;
+    public void setFilename(String filename) {
+        this.filename = filename;
     }
-
-    public int getDownloadStatusText() {
-        return downloadStatus;
-    }
-
-    public long getBatchId() {
-        return batchId;
-    }
-
-    public boolean isPaused() {
-        return downloadStatus == DownloadManager.STATUS_PAUSED;
-    }
-
 }
