@@ -14,9 +14,8 @@ public class Constants {
     public static final String JSON_POST_UTF8 = "UTF-8";
 
     public static final String HTTPS = "https://";
-    public static final String HTTP = "http://";
     public static final String DOMAIN = "oaza.tv/";
-    public static final String API = HTTP + DOMAIN + "api/v2/";
+    public static final String API = HTTPS + DOMAIN + "api/v2/";
     public static final String API_GET_ARCHIVE = API + "archive/";
     public static final String API_GET_PHOTO_ALBUM = API + "album/";
     public static final String API_SEARCH = API + "search/";
@@ -31,17 +30,6 @@ public class Constants {
     public static final String API_GET_SONGS = API + "songs/";
     public static final String SIMILAR = "/similar";
     public static final String VIEW = "/view";
-
-    public static final String JSON_POST_NICK = "nick";
-    public static final String JSON_POST_EMAIL = "email";
-    public static final String JSON_POST_PASS = "password";
-    public static final String JSON_POST_NAME = "name";
-    public static final String JSON_POST_TOKEN = "token";
-    public static final String JSON_POST_IMAGE = "image";
-    public static final String JSON_POST_TEXT = "text";
-    public static final String JSON_POST_FEEDBACK = "feedback";
-    public static final String JSON_POST_PURCHASE_TOKEN = "purchase_token";
-    public static final String JSON_POST_LIKE = "like";
 
     public static final String GOOGLE_PLAY_APP_LINK = "https://play.google.com/store/apps/details?id=com.chaemil.hgms";
     public static final String LIVESTREAM_LINK = DOMAIN + "/live-stream/";
